+++ src/app/layout/Header.css
@@ -1,4 +1,4 @@
-@import './variables.css';
+@import '../utils/variables.css';
 
 .header {
   display: flex;
