+++ src/home/About.css
@@ -1,4 +1,4 @@
-@import '../app/variables.css';
+@import '../app/utils/variables.css';
 
 #about {
   padding: 3.75rem 0 9%;
