+++ src/static.jsx
@@ -3,11 +3,11 @@ import React from 'react';
 import { StaticRouter } from 'react-router-dom';
 
 import renderHtml from '../public/index.jsx';
-import App from './app/App.jsx';
+import Root from './app/Root.jsx';
 
 export default function render({ path: location, webpackStats }) {
   return renderHtml(
     webpackStats,
-    <App ContentTag="body" Router={StaticRouter} routerProps={{ location }} />,
+    <Root ContentTag="body" Router={StaticRouter} routerProps={{ location }} />,
   );
 }
